requirements.txt
pyspark==3.5.0
pandas==2.1.4
pyarrow==14.0.2
apache-airflow==2.8.1
apache-airflow-providers-snowflake==5.2.1
apache-airflow-providers-apache-spark==4.7.0
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
great-expectations==0.18.8
streamlit==1.29.0
plotly==5.18.0
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
kafka-python==2.0.2
boto3==1.34.14
pytest==7.4.4
pytest-spark==0.6.0
